# packages/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kobai",
    "version": "0.0.0",
    "description": "kobai's HTTP surface. Two authenticated surfaces: `/admin`, behind a Merchant session, and `/store`, behind an API key (ADR-0020). `/health` is open, and is the only route that answers before migrations have applied.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "merchantSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Merchant session token, from `POST /admin/session`. Opens `/admin` and nothing else."
      },
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "kobai_pk_… | kobai_sk_…",
        "description": "An API key, from `POST /admin/api-keys`. Opens `/store` and nothing else. The prefix says which kind it is, with no lookup: `kobai_pk_` is publishable and safe in a browser, `kobai_sk_` is secret and is not."
      }
    },
    "schemas": {
...
      "MigrationState": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "pending"
                ]
              }
            },
            "required": [
              "status"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "running"
                ]
              }
            },
            "required": [
              "status"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "applied"
                ]
              },
              "sets": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AppliedMigrationSet"
                }
              }
            },
            "required": [
              "status",
              "sets"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "failed"
                ]
              },
              "set": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "status",
              "set",
              "message"
            ]
          }
        ]
      },
      "AppliedMigrationSet": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "migrationsTable": {
            "type": "string"
          },
          "migrationsSchema": {
            "type": "string"
          },
          "applied": {
            "type": "integer",
            "description": "Rows in this set's tracking table."
          }
        },
        "required": [
          "name",
          "migrationsTable",
          "migrationsSchema",
          "applied"
        ]
      },
...
      "RoleSummary": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "What this Role may do. A deployment may hold a permission this build of Core has never heard of."
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "MerchantIdentity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "Refusal": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, in prose."
          },
          "reason": {
            "type": "string",
            "description": "Machine-readable. Branch on this."
          }
        },
        "required": [
          "error",
          "reason"
        ]
      },
      "SessionRefusal": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "session-missing",
              "session-malformed",
              "session-unknown",
              "session-expired"
            ]
          }
        },
        "required": [
          "error",
          "reason"
        ]
      },
      "PermissionDenied": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "permission-denied"
            ]
          },
          "required": {
            "type": "string",
            "description": "The permission the Role does not hold."
          }
        },
        "required": [
          "error",
          "reason",
          "required"
        ]
      },
      "ServerError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Unavailable": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "booting",
              "error"
            ]
          },
          "migrations": {
            "$ref": "#/components/schemas/MigrationState"
          }
        },
        "required": [
          "error",
          "status",
          "migrations"
        ]
      },
...
      "IssuedSession": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Session"
          },
          {
            "type": "object",
            "properties": {
              "token": {
                "type": "string"
              }
            },
            "required": [
              "token"
            ]
          }
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "merchant": {
            "$ref": "#/components/schemas/MerchantIdentity"
          },
          "role": {
            "$ref": "#/components/schemas/RoleSummary"
          }
        },
        "required": [
          "expiresAt",
          "merchant",
          "role"
        ]
      },
      "InvalidCredentials": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "invalid-credentials"
            ]
          }
        },
        "required": [
          "error",
          "reason"
        ]
      },
      "SignInRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "Store": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "defaultCurrency": {
            "type": "string",
            "description": "ISO 4217, upper case."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {},
            "description": "Unindexed, untyped JSON owned by the Merchant and the Project."
          }
        },
        "required": [
          "name",
          "defaultCurrency",
          "metadata"
        ]
      },
...
      "IssuedApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ApiKeyKind"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "key": {
            "type": "string",
            "description": "The value itself. Shown at creation and unrecoverable afterwards."
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "createdAt",
          "key"
        ]
      },
      "ApiKeyKind": {
        "type": "string",
        "enum": [
          "publishable",
          "secret"
        ]
      },
      "CreateApiKeyRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "How a Merchant tells one key from another when revoking."
          },
          "kind": {
            "$ref": "#/components/schemas/ApiKeyKind"
          }
        },
        "required": [
          "name",
          "kind"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/admin/session": {
      "post": {
        "summary": "Sign in",
        "description": "An unknown address and a wrong password are answered identically, and in the same time. Distinguishing them would turn this into a way to ask who works here.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The session, with the token to present on every later request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedSession"
                }
              }
            }
          },
          "400": {
            "description": "The request does not fit this endpoint's schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "401": {
            "description": "Those credentials are not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidCredentials"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Who the caller is",
        "description": "What the Admin asks first after a page load: who am I, and what may I do.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "The caller's Merchant and Role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Sign out",
        "description": "The row goes, so the token stops working on the very next request.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "responses": {
          "204": {
            "description": "Signed out."
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/admin/store": {
      "get": {
        "summary": "Read the Store",
        "description": "One deployment is one Store (ADR-0005), so this takes no identifier.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "The Store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Store"
                }
              }
            }
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "403": {
            "description": "The Merchant's Role does not hold the permission this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDenied"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api-keys": {
      "post": {
        "summary": "Mint an API key",
        "description": "The value is in this response and in no other, ever — only a digest is stored. `kobai_pk_…` is publishable and `kobai_sk_…` is secret, so the kind is readable off the value itself.",
        "security": [
          {
            "merchantSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The key, shown once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedApiKey"
                }
              }
            }
          },
          "400": {
            "description": "The request does not fit this endpoint's schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "401": {
            "description": "No live Merchant session was presented.",
            "headers": {
              "www-authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ],
                  "description": "The scheme the request failed to satisfy."
                },
                "required": true,
                "description": "The scheme the request failed to satisfy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefusal"
                }
              }
            }
          },
          "403": {
            "description": "The Merchant's Role does not hold the permission this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDenied"
                }
              }
            }
          },
          "500": {
            "description": "Something failed inside kobai.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "503": {
            "description": "Migrations have not applied, so nothing but `/health` is served yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}